test: split monolithic parser and plc2plc test files into focused modules
Feature PRs repeatedly conflict in the large, append-only test modules
`parser/src/tests.rs` (~3060 lines) and `plc2plc/src/tests.rs` (~678
lines): every PR adds to the same shared `use { … }` block and appends
`#[test]` fns on the same trailing lines, so Git conflicts are effectively
guaranteed.

Adopt the structure the codegen `tests/it/` suite already uses. Each
monolith becomes a `tests/` directory: shared imports and helpers move to a
`common` submodule (re-exported `pub(crate)`), and every feature area gets
its own file (`case.rs`, `enums.rs`, `reference_to.rs`, `duration.rs`, …)
registered with one `mod` line in `tests/mod.rs`. Adding tests for a new
feature is now a new file plus one `mod` line, so branches no longer collide
on shared lines.

This is a pure move — no test body is edited; the set of test and helper
function names is identical before and after, and the full suite passes
(parser 292 + 3 ignored, plc2plc 48). Also gate plc2plc's `mod tests` behind
`#[cfg(test)]` to match parser (the split surfaced that its dev-dependency
imports were compiled in non-test builds).

Documents the convention in the syntax-support guide and refreshes the
now-stale `tests.rs` paths in the living design docs.
