--- compiler/parser/src/tests.rs
@@ -1,3060 +1,0 @@
-//! Tests of parser.
-#[cfg(test)]
-mod test {
-    use dsl::common::{
-        next_block_id, ArrayElementType, ConstantKind, DataTypeDeclarationKind,
-        DeclarationQualifier, EnumeratedSpecificationInit, EnumerationDeclaration,
-        FunctionBlockBodyKind, FunctionBlockDeclaration, FunctionDeclaration, FunctionReturnType,
-        InitialValueAssignmentKind, Library, LibraryElementKind, ProgramDeclaration, RealLiteral,
-        ReferenceTarget, SimpleInitializer, SpecificationKind, TypeName, TypeReference, VarDecl,
-        VariableIdentifier, VariableType,
-    };
-    use dsl::configuration::{
-        ConfigurationDeclaration, DataSourceKind, ProgramConfiguration, ResourceDeclaration,
-        TaskConfiguration,
-    };
-    use dsl::core::{FileId, Id, SourceSpan};
-    use dsl::diagnostic::Diagnostic;
-    use dsl::sfc::{ActionAssociation, ActionQualifier, ElementKind, Network, Step};
-    use dsl::textual::*;
-    use dsl::time::*;
-    use ironplc_test::cast;
-    use ironplc_test::read_shared_resource;
-    use rstest::rstest;
-    use time::Duration;
-
-    use crate::options::{CompilerOptions, Dialect};
-    use crate::parse_program;
-
-    pub fn parse_resource(name: &'static str) -> Result<Library, Diagnostic> {
-        let source = read_shared_resource(name);
-        parse_program(&source, &FileId::default(), &CompilerOptions::default())
-    }
-
-    fn parse_text(source: &'static str) -> Library {
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_ok());
-        result.unwrap()
-    }
-
-    #[test]
-    fn parse_variable_declarations() {
-        let res = parse_resource("var_decl.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_inout_variable_declarations() {
-        let res = parse_resource("inout_var_decl.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_input_variable_declarations() {
-        let res = parse_resource("input_var_decl.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_strings() {
-        let res = parse_resource("strings.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_type_decl() {
-        let res = parse_resource("type_decl.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_textual() {
-        let res = parse_resource("textual.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_conditional() {
-        let res = parse_resource("conditional.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_oscat() {
-        // OSCAT files have a header that as far as I can tell is not valid
-        // but it is common.
-        let res = parse_resource("oscat.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_expressions() {
-        let res = parse_resource("expressions.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_array() {
-        let res = parse_resource("array.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_nested() {
-        let res: Result<Library, Diagnostic> = parse_resource("nested.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_configuration() {
-        let res: Result<Library, Diagnostic> = parse_resource("configuration.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_program_then_ok() {
-        let res: Result<Library, Diagnostic> = parse_resource("program.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_if_then_ok() {
-        let source = ironplc_test::read_shared_resource("if.st");
-        let options = CompilerOptions {
-            allow_missing_semicolon: true,
-            ..Default::default()
-        };
-        let res = parse_program(&source, &FileId::default(), &options);
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_program_when_has_comment_then_ok() {
-        let source = "
-        TYPE
-        (* A comment *)
-            CUSTOM_STRUCT : STRUCT 
-                NAME: BOOL;
-            END_STRUCT;
-        END_TYPE";
-
-        let res = parse_text(source);
-        assert_eq!(1, res.elements.len());
-    }
-
-    #[test]
-    fn parse_program_when_back_to_back_comments_then_ok() {
-        let program = "
-        TYPE
-        (* A comment *)(* A comment *)
-           CUSTOM_STRUCT : STRUCT 
-             NAME: BOOL;
-           END_STRUCT;
-        END_TYPE";
-
-        parse_text(program);
-    }
-
-    #[test]
-    fn parse_program_when_right_parent_in_comment_then_ok() {
-        let program = "
-        TYPE
-        (* A comment) *)(* A comment *)
-           CUSTOM_STRUCT : STRUCT 
-             NAME: BOOL;
-           END_STRUCT;
-        END_TYPE";
-
-        parse_text(program);
-    }
-
-    #[test]
-    fn parse_program_when_comment_not_closed_then_err() {
-        let program = "
-        TYPE
-        (* A comment
-            CUSTOM_STRUCT : STRUCT
-                NAME: BOOL;
-            END_STRUCT;
-        END_TYPE";
-
-        let res = parse_program(program, &FileId::default(), &CompilerOptions::default());
-        assert!(res.is_err());
-    }
-
-    #[test]
-    fn parse_program_when_bad_name_then_err() {
-        let program = "
-        TYPE
-            CUSTOM_STRUCT : STRUCT& 
-                NAME: BOOL;
-            END_STRUCT;
-        END_TYPE";
-
-        let res = parse_program(program, &FileId::default(), &CompilerOptions::default());
-        assert!(res.is_err());
-
-        let err = res.unwrap_err();
-        assert_eq!("Syntax error".to_owned(), err.description());
-        assert_eq!("Expected ' ' (space) | '\\t' (tab) | '(* ... *)' (comment) | '\\n' (new line) | '{ ... }' (pragma) | (identifier). Found text '&' that matched token 'AND' | '&'".to_owned(), err.primary.message);
-    }
-
-    #[test]
-    fn parse_program_when_not_valid_top_item_then_err() {
-        let program = "ACTION
-        END_ACTION";
-
-        let res = parse_program(program, &FileId::default(), &CompilerOptions::default());
-        assert!(res.is_err());
-
-        let err = res.unwrap_err();
-        assert_eq!("Syntax error".to_owned(), err.description());
-        assert_eq!("Expected ' ' (space) | '\\t' (tab) | '(* ... *)' (comment) | 'CONFIGURATION' | 'FUNCTION' | 'FUNCTION_BLOCK' | 'PROGRAM' | 'TYPE' | 'VAR_GLOBAL' | '\\n' (new line) | '{ ... }' (pragma). Found text 'ACTION' that matched token 'ACTION'".to_owned(), err.primary.message);
-    }
-
-    #[test]
-    fn parse_program_when_complex_bit_string_then_ok() {
-        let program = "
-FUNCTION fun:DWORD
-
-VAR_IN_OUT
-    VAR1: INT;
-END_VAR
-
-VAR1 := DWORD#16#0000FFFF;
-
-END_FUNCTION";
-
-        parse_text(program);
-    }
-
-    #[test]
-    fn parse_program_when_real_then_ok() {
-        let program = "
-FUNCTION fun:DWORD
-
-VAR
-    InputsNumber : REAL := -5.0E-1;
-END_VAR
-
-fun := InputsNumber;
-
-END_FUNCTION";
-        let res = parse_text(program);
-
-        let expected = new_library(LibraryElementKind::FunctionDeclaration(
-            FunctionDeclaration {
-                name: Id::from("fun"),
-                return_type: FunctionReturnType::Named(TypeName::from("DWORD")),
-                variables: vec![VarDecl {
-                    identifier: VariableIdentifier::new_symbol("InputsNumber"),
-                    var_type: VariableType::Var,
-                    qualifier: DeclarationQualifier::Unspecified,
-                    initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                        type_name: TypeName::from("REAL"),
-                        initial_value: Some(ConstantKind::RealLiteral(RealLiteral {
-                            value: -0.5,
-                            data_type: None,
-                        })),
-                    }),
-                    block: next_block_id(),
-                }],
-                edge_variables: vec![],
-                body: vec![StmtKind::simple_assignment("fun", "InputsNumber")],
-            },
-        ));
-        assert_eq!(res, expected);
-    }
-
-    #[test]
-    fn parse_program_when_real_scientific_no_decimal_then_ok() {
-        let program = "
-FUNCTION fun:DWORD
-
-VAR
-    InputsNumber : REAL := 2E-3;
-END_VAR
-
-fun := InputsNumber;
-
-END_FUNCTION";
-        let res = parse_text(program);
-
-        let expected = new_library(LibraryElementKind::FunctionDeclaration(
-            FunctionDeclaration {
-                name: Id::from("fun"),
-                return_type: FunctionReturnType::Named(TypeName::from("DWORD")),
-                variables: vec![VarDecl {
-                    identifier: VariableIdentifier::new_symbol("InputsNumber"),
-                    var_type: VariableType::Var,
-                    qualifier: DeclarationQualifier::Unspecified,
-                    initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                        type_name: TypeName::from("REAL"),
-                        initial_value: Some(ConstantKind::RealLiteral(RealLiteral {
-                            value: 0.002,
-                            data_type: None,
-                        })),
-                    }),
-                    block: next_block_id(),
-                }],
-                edge_variables: vec![],
-                body: vec![StmtKind::simple_assignment("fun", "InputsNumber")],
-            },
-        ));
-        assert_eq!(res, expected);
-    }
-
-    #[test]
-    fn parse_program_when_real_scientific_positive_exponent_then_ok() {
-        let program = "
-FUNCTION fun:DWORD
-
-VAR
-    InputsNumber : REAL := 1.5E+2;
-END_VAR
-
-fun := InputsNumber;
-
-END_FUNCTION";
-        let res = parse_text(program);
-
-        let expected = new_library(LibraryElementKind::FunctionDeclaration(
-            FunctionDeclaration {
-                name: Id::from("fun"),
-                return_type: FunctionReturnType::Named(TypeName::from("DWORD")),
-                variables: vec![VarDecl {
-                    identifier: VariableIdentifier::new_symbol("InputsNumber"),
-                    var_type: VariableType::Var,
-                    qualifier: DeclarationQualifier::Unspecified,
-                    initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                        type_name: TypeName::from("REAL"),
-                        initial_value: Some(ConstantKind::RealLiteral(RealLiteral {
-                            value: 150.0,
-                            data_type: None,
-                        })),
-                    }),
-                    block: next_block_id(),
-                }],
-                edge_variables: vec![],
-                body: vec![StmtKind::simple_assignment("fun", "InputsNumber")],
-            },
-        ));
-        assert_eq!(res, expected);
-    }
-
-    #[test]
-    fn parse_program_when_fixed_point_duration_then_ok() {
-        let program = "
-FUNCTION fun:TIME
-
-VAR
-    tv : TIME := t#1.2s;
-END_VAR
-
-fun := tv;
-
-END_FUNCTION";
-        let actual = parse_text(program);
-
-        let expected = new_library(LibraryElementKind::FunctionDeclaration(
-            FunctionDeclaration {
-                name: Id::from("fun"),
-                return_type: FunctionReturnType::Named(TypeName::from("TIME")),
-                variables: vec![VarDecl {
-                    identifier: VariableIdentifier::new_symbol("tv"),
-                    var_type: VariableType::Var,
-                    qualifier: DeclarationQualifier::Unspecified,
-                    initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                        type_name: TypeName::from("TIME"),
-                        initial_value: Some(ConstantKind::Duration(DurationLiteral {
-                            interval: Duration::milliseconds(1200),
-                            span: SourceSpan::default(),
-                        })),
-                    }),
-                    block: next_block_id(),
-                }],
-                edge_variables: vec![],
-                body: vec![StmtKind::simple_assignment("fun", "tv")],
-            },
-        ));
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn parse_when_first_steps_function_block_counter_fbd_then_builds_structure() {
-        let actual = parse_resource("first_steps_function_block_counter_fbd.st").unwrap();
-        let expected = new_library(LibraryElementKind::FunctionBlockDeclaration(
-            FunctionBlockDeclaration {
-                name: TypeName::from("CounterFBD"),
-                variables: vec![
-                    VarDecl::simple("Reset", "BOOL").with_type(VariableType::Input),
-                    VarDecl::simple("OUT", "INT").with_type(VariableType::Output),
-                    VarDecl::simple("Cnt", "INT"),
-                    VarDecl {
-                        identifier: VariableIdentifier::new_symbol("ResetCounterValue"),
-                        var_type: VariableType::External,
-                        qualifier: DeclarationQualifier::Constant,
-                        initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                            type_name: TypeName::from("INT"),
-                            initial_value: None,
-                        }),
-                        block: next_block_id(),
-                    },
-                    VarDecl::simple("_TMP_ADD4_OUT", "INT"),
-                    VarDecl::simple("_TMP_SEL7_OUT", "INT"),
-                ],
-                edge_variables: vec![],
-                body: FunctionBlockBodyKind::stmts(vec![
-                    StmtKind::simple_assignment("Cnt", "_TMP_SEL7_OUT"),
-                    StmtKind::simple_assignment("OUT", "Cnt"),
-                ]),
-                span: SourceSpan::default(),
-            },
-        ));
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn parse_when_first_steps_func_avg_val_then_builds_structure() {
-        let actual = parse_resource("first_steps_func_avg_val.st").unwrap();
-        let expected = new_library(LibraryElementKind::FunctionDeclaration(
-            FunctionDeclaration {
-                name: Id::from("AverageVal"),
-                return_type: FunctionReturnType::Named(TypeName::from("REAL")),
-                variables: vec![
-                    VarDecl::simple("Cnt1", "INT").with_type(VariableType::Input),
-                    VarDecl::simple("Cnt2", "INT").with_type(VariableType::Input),
-                    VarDecl::simple("Cnt3", "INT").with_type(VariableType::Input),
-                    VarDecl::simple("Cnt4", "INT").with_type(VariableType::Input),
-                    VarDecl::simple("Cnt5", "INT").with_type(VariableType::Input),
-                    VarDecl {
-                        identifier: VariableIdentifier::new_symbol("InputsNumber"),
-                        var_type: VariableType::Var,
-                        qualifier: DeclarationQualifier::Unspecified,
-                        initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                            type_name: TypeName::from("REAL"),
-                            initial_value: Some(ConstantKind::RealLiteral(RealLiteral {
-                                value: 5.1,
-                                data_type: None,
-                            })),
-                        }),
-                        block: next_block_id(),
-                    },
-                ],
-                edge_variables: vec![],
-                body: vec![StmtKind::assignment(
-                    Variable::named("AverageVal"),
-                    ExprKind::binary(
-                        Operator::Div,
-                        ExprKind::Function(Function {
-                            name: Id::from("INT_TO_REAL"),
-                            param_assignment: vec![ParamAssignmentKind::positional(
-                                ExprKind::binary(
-                                    Operator::Add,
-                                    ExprKind::binary(
-                                        Operator::Add,
-                                        ExprKind::binary(
-                                            Operator::Add,
-                                            ExprKind::binary(
-                                                Operator::Add,
-                                                ExprKind::late_bound("Cnt1"),
-                                                ExprKind::late_bound("Cnt2"),
-                                            ),
-                                            ExprKind::late_bound("Cnt3"),
-                                        ),
-                                        ExprKind::late_bound("Cnt4"),
-                                    ),
-                                    ExprKind::late_bound("Cnt5"),
-                                ),
-                            )],
-                        }),
-                        ExprKind::late_bound("InputsNumber"),
-                    ),
-                )],
-            },
-        ));
-        assert_eq!(actual, expected)
-    }
-
-    #[test]
-    fn parse_when_first_steps_program_declaration_then_builds_structure() {
-        let actual = parse_resource("first_steps_program.st").unwrap();
-        let expected = new_library(LibraryElementKind::ProgramDeclaration(ProgramDeclaration {
-            name: Id::from("plc_prg"),
-            variables: vec![
-                VarDecl::simple("Reset", "BOOL").with_type(VariableType::Input),
-                VarDecl::simple("Cnt1", "INT").with_type(VariableType::Output),
-                VarDecl::simple("Cnt2", "INT").with_type(VariableType::Output),
-                VarDecl::simple("Cnt3", "INT").with_type(VariableType::Output),
-                VarDecl::simple("Cnt4", "INT").with_type(VariableType::Output),
-                VarDecl::simple("Cnt5", "INT").with_type(VariableType::Output),
-                VarDecl::late_bound("CounterST0", "CounterST"),
-                VarDecl::late_bound("CounterFBD0", "CounterFBD"),
-                VarDecl::late_bound("CounterSFC0", "CounterSFC"),
-                VarDecl::late_bound("CounterIL0", "CounterIL"),
-                VarDecl::late_bound("CounterLD0", "CounterLD"),
-                VarDecl::simple("AVCnt", "REAL"),
-                VarDecl::simple("_TMP_AverageVal17_OUT", "REAL"),
-            ],
-            access_variables: vec![],
-            body: FunctionBlockBodyKind::stmts(vec![
-                StmtKind::fb_call_mapped("CounterST0", vec![("Reset", "Reset")]),
-                StmtKind::structured_assignment("Cnt1", "CounterST0", "OUT"),
-                StmtKind::fb_assign(
-                    "AverageVal",
-                    vec!["Cnt1", "Cnt2", "Cnt3", "Cnt4", "Cnt5"],
-                    "_TMP_AverageVal17_OUT",
-                ),
-                StmtKind::simple_assignment("AVCnt", "_TMP_AverageVal17_OUT"),
-                StmtKind::fb_call_mapped("CounterFBD0", vec![("Reset", "Reset")]),
-                StmtKind::structured_assignment("Cnt2", "CounterFBD0", "OUT"),
-                StmtKind::fb_call_mapped("CounterSFC0", vec![("Reset", "Reset")]),
-                StmtKind::structured_assignment("Cnt3", "CounterSFC0", "OUT"),
-                StmtKind::fb_call_mapped("CounterIL0", vec![("Reset", "Reset")]),
-                StmtKind::structured_assignment("Cnt4", "CounterIL0", "OUT"),
-                StmtKind::fb_call_mapped("CounterLD0", vec![("Reset", "Reset")]),
-                StmtKind::structured_assignment("Cnt5", "CounterLD0", "Out"),
-            ]),
-        }));
-        assert_eq!(actual, expected)
-    }
-
-    #[test]
-    fn parse_when_first_steps_configuration_then_builds_structure() {
-        let actual = parse_resource("first_steps_configuration.st").unwrap();
-        let expected = new_library(LibraryElementKind::ConfigurationDeclaration(
-            ConfigurationDeclaration {
-                name: Id::from("config"),
-                global_var: vec![VarDecl {
-                    identifier: VariableIdentifier::new_symbol("ResetCounterValue"),
-                    var_type: VariableType::Global,
-                    qualifier: DeclarationQualifier::Constant,
-                    initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                        type_name: TypeName::from("INT"),
-                        initial_value: Some(ConstantKind::integer_literal("17").unwrap()),
-                    }),
-                    block: next_block_id(),
-                }],
-                resource_decl: vec![ResourceDeclaration {
-                    name: Id::from("resource1"),
-                    resource: Id::from("PLC"),
-                    global_vars: vec![],
-                    tasks: vec![TaskConfiguration {
-                        name: Id::from("plc_task"),
-                        priority: 1,
-                        interval: Option::Some(DurationLiteral {
-                            span: SourceSpan::default(),
-                            interval: Duration::new(0, 100_000_000),
-                        }),
-                        single: None,
-                    }],
-                    programs: vec![ProgramConfiguration {
-                        name: Id::from("plc_task_instance"),
-                        storage: None,
-                        task_name: Option::Some(Id::from("plc_task")),
-                        type_name: Id::from("plc_prg"),
-                        fb_tasks: vec![],
-                        sources: vec![],
-                        sinks: vec![],
-                    }],
-                }],
-                fb_inits: vec![],
-                located_var_inits: vec![],
-            },
-        ));
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn parse_when_first_steps_function_block_logger_then_test_apply_when_names_correct_then_passes()
-    {
-        let actual = parse_resource("first_steps_function_block_logger.st").unwrap();
-        let expected = new_library(LibraryElementKind::FunctionBlockDeclaration(
-            FunctionBlockDeclaration {
-                name: TypeName::from("LOGGER"),
-                variables: vec![
-                    VarDecl::simple("TRIG", "BOOL").with_type(VariableType::Input),
-                    VarDecl::string(
-                        "MSG",
-                        VariableType::Input,
-                        DeclarationQualifier::Unspecified,
-                    ),
-                    VarDecl::enumerated("LEVEL", "LOGLEVEL", "INFO").with_type(VariableType::Input),
-                    VarDecl::simple("TRIG0", "BOOL"),
-                ],
-                edge_variables: vec![],
-                body: FunctionBlockBodyKind::stmts(vec![
-                    StmtKind::if_then(
-                        ExprKind::compare(
-                            CompareOp::And,
-                            ExprKind::late_bound("TRIG"),
-                            ExprKind::unary(UnaryOp::Not, ExprKind::late_bound("TRIG0")),
-                        ),
-                        vec![],
-                    ),
-                    StmtKind::assignment(Variable::named("TRIG0"), ExprKind::late_bound("TRIG")),
-                ]),
-                span: SourceSpan::default(),
-            },
-        ));
-
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn parse_when_first_steps_function_block_counter_sfc_then_builds_structure() {
-        let actual = parse_resource("first_steps_function_block_counter_sfc.st").unwrap();
-        let expected = new_library(LibraryElementKind::FunctionBlockDeclaration(
-            FunctionBlockDeclaration {
-                name: TypeName::from("CounterSFC"),
-                variables: vec![
-                    VarDecl::simple("Reset", "BOOL").with_type(VariableType::Input),
-                    VarDecl::simple("OUT", "INT").with_type(VariableType::Output),
-                    VarDecl::simple("Cnt", "INT"),
-                    VarDecl {
-                        identifier: VariableIdentifier::new_symbol("ResetCounterValue"),
-                        var_type: VariableType::External,
-                        qualifier: DeclarationQualifier::Constant,
-                        initializer: InitialValueAssignmentKind::Simple(SimpleInitializer {
-                            type_name: TypeName::from("INT"),
-                            initial_value: None,
-                        }),
-                        block: next_block_id(),
-                    },
-                ],
-                edge_variables: vec![],
-                body: FunctionBlockBodyKind::sfc(vec![Network {
-                    initial_step: Step {
-                        name: Id::from("Start"),
-                        action_associations: vec![],
-                    },
-                    elements: vec![
-                        ElementKind::transition(
-                            "Start",
-                            "ResetCounter",
-                            ExprKind::late_bound("Reset"),
-                        ),
-                        ElementKind::step(
-                            Id::from("ResetCounter"),
-                            vec![
-                                ActionAssociation::new(
-                                    "RESETCOUNTER_INLINE1",
-                                    Some(ActionQualifier::N),
-                                ),
-                                ActionAssociation::new(
-                                    "RESETCOUNTER_INLINE2",
-                                    Some(ActionQualifier::N),
-                                ),
-                            ],
-                        ),
-                        ElementKind::action(
-                            "RESETCOUNTER_INLINE1",
-                            vec![StmtKind::simple_assignment("Cnt", "ResetCounterValue")],
-                        ),
-                        ElementKind::action(
-                            "RESETCOUNTER_INLINE2",
-                            vec![StmtKind::simple_assignment("OUT", "Cnt")],
-                        ),
-                        ElementKind::transition(
-                            "ResetCounter",
-                            "Start",
-                            ExprKind::unary(UnaryOp::Not, ExprKind::late_bound("Reset")),
-                        ),
-                        ElementKind::transition(
-                            "Start",
-                            "Count",
-                            ExprKind::unary(UnaryOp::Not, ExprKind::late_bound("Reset")),
-                        ),
-                        ElementKind::step(
-                            Id::from("Count"),
-                            vec![
-                                ActionAssociation::new("COUNT_INLINE3", Some(ActionQualifier::N)),
-                                ActionAssociation::new("COUNT_INLINE4", Some(ActionQualifier::N)),
-                            ],
-                        ),
-                        ElementKind::action(
-                            "COUNT_INLINE3",
-                            vec![StmtKind::assignment(
-                                Variable::named("Cnt"),
-                                ExprKind::binary(
-                                    Operator::Add,
-                                    ExprKind::late_bound("Cnt"),
-                                    ExprKind::integer_literal("1"),
-                                ),
-                            )],
-                        ),
-                        ElementKind::action(
-                            "COUNT_INLINE4",
-                            vec![StmtKind::simple_assignment("OUT", "Cnt")],
-                        ),
-                        ElementKind::transition("Count", "Start", ExprKind::late_bound("Reset")),
-                    ],
-                }]),
-                span: SourceSpan::default(),
-            },
-        ));
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn parse_when_first_steps_data_type_decl_then_builds_structure() {
-        let actual = parse_resource("first_steps_data_type_decl.st").unwrap();
-        let expected = new_library(LibraryElementKind::DataTypeDeclaration(
-            DataTypeDeclarationKind::Enumeration(EnumerationDeclaration {
-                type_name: TypeName::from("LOGLEVEL"),
-                spec_init: EnumeratedSpecificationInit::values_and_default(
-                    vec!["CRITICAL", "WARNING", "INFO", "DEBUG"],
-                    "INFO",
-                ),
-            }),
-        ));
-        assert_eq!(actual, expected)
-    }
-
-    #[cfg(test)]
-    pub fn new_library(element: LibraryElementKind) -> Library {
-        Library {
-            elements: vec![element],
-        }
-    }
-
-    #[test]
-    fn parse_when_enumerated_value_is_reserved_keyword_then_parses() {
-        // enumerated_value() used to require a bare identifier(), which
-        // rejects ON/STEP/R_EDGE/F_EDGE (reserved tokens) even though
-        // variable_identifier() already carves out exactly this set for
-        // VAR declarations (see #300, "Feature/reserved variables") --
-        // real-world enums commonly use these as ordinary member names
-        // (e.g. `(off, on)` for a blink state).
-        let source = "TYPE E_Test : (on, off, step, r_edge, f_edge, normal_value); END_TYPE";
-        let library = parse_text(source);
-        let decl = cast!(
-            &library.elements[0],
-            LibraryElementKind::DataTypeDeclaration
-        );
-        let enumeration = cast!(decl, DataTypeDeclarationKind::Enumeration);
-        let inline = cast!(&enumeration.spec_init.spec, SpecificationKind::Inline);
-        let values: Vec<&String> = inline.values.iter().map(|v| v.value.original()).collect();
-        assert_eq!(
-            values,
-            vec!["on", "off", "step", "r_edge", "f_edge", "normal_value"]
-        );
-    }
-
-    #[test]
-    fn parse_when_function_with_any_num_return_type_then_parses() {
-        let lib = parse_text(
-            "FUNCTION ABS : ANY_NUM
-            VAR_INPUT
-                IN : ANY_NUM;
-            END_VAR
-            ABS := IN;
-            END_FUNCTION",
-        );
-
-        assert_eq!(lib.elements.len(), 1);
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        assert_eq!(
-            func.return_type,
-            FunctionReturnType::Named(TypeName::from("ANY_NUM"))
-        );
-    }
-
-    #[test]
-    fn parse_when_variable_with_any_int_type_then_parses() {
-        let lib = parse_text(
-            "FUNCTION TEST : INT
-            VAR_INPUT
-                val : ANY_INT;
-            END_VAR
-            TEST := 0;
-            END_FUNCTION",
-        );
-
-        assert_eq!(lib.elements.len(), 1);
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let type_ref = func.variables[0].type_name();
-        assert_eq!(type_ref, TypeReference::Named(TypeName::from("ANY_INT")));
-    }
-
-    #[test]
-    fn parse_when_all_generic_type_names_then_parses() {
-        // Test all generic type names can be used as type references
-        let generic_types = [
-            "ANY",
-            "ANY_DERIVED",
-            "ANY_ELEMENTARY",
-            "ANY_MAGNITUDE",
-            "ANY_NUM",
-            "ANY_REAL",
-            "ANY_INT",
-            "ANY_BIT",
-            "ANY_STRING",
-            "ANY_DATE",
-        ];
-
-        for generic_type in generic_types {
-            let source = format!(
-                "FUNCTION TEST : {}
-                VAR_INPUT
-                    val : {};
-                END_VAR
-                TEST := val;
-                END_FUNCTION",
-                generic_type, generic_type
-            );
-
-            let result = parse_program(&source, &FileId::default(), &CompilerOptions::default());
-            assert!(
-                result.is_ok(),
-                "Failed to parse generic type: {}",
-                generic_type
-            );
-
-            let lib = result.unwrap();
-            let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-            assert_eq!(
-                func.return_type,
-                FunctionReturnType::Named(TypeName::from(generic_type)),
-                "Return type mismatch for {}",
-                generic_type
-            );
-        }
-    }
-
-    #[test]
-    fn parse_when_function_with_string_length_return_type_then_parses() {
-        let lib = parse_text(
-            "FUNCTION my_func : STRING[255]
-            VAR_INPUT
-                x : INT;
-            END_VAR
-            my_func := 'hello';
-            END_FUNCTION",
-        );
-
-        assert_eq!(lib.elements.len(), 1);
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let spec = cast!(&func.return_type, FunctionReturnType::String);
-        assert_eq!(spec.width, dsl::common::StringType::String);
-        assert!(spec.length.is_some());
-    }
-
-    #[test]
-    fn parse_when_function_with_wstring_length_return_type_then_parses() {
-        let lib = parse_text(
-            "FUNCTION my_func : WSTRING[100]
-            VAR_INPUT
-                x : INT;
-            END_VAR
-            my_func := 'hello';
-            END_FUNCTION",
-        );
-
-        assert_eq!(lib.elements.len(), 1);
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let spec = cast!(&func.return_type, FunctionReturnType::WString);
-        assert_eq!(spec.width, dsl::common::StringType::WString);
-        assert!(spec.length.is_some());
-    }
-
-    #[test]
-    fn parse_when_function_with_bare_string_return_type_then_parses() {
-        let lib = parse_text(
-            "FUNCTION my_func : STRING
-            VAR_INPUT
-                x : INT;
-            END_VAR
-            my_func := 'hello';
-            END_FUNCTION",
-        );
-
-        assert_eq!(lib.elements.len(), 1);
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let spec = cast!(&func.return_type, FunctionReturnType::String);
-        assert_eq!(spec.width, dsl::common::StringType::String);
-        assert!(spec.length.is_none());
-    }
-
-    #[test]
-    fn parse_when_function_var_with_string_length_then_parses() {
-        let _lib = parse_text(
-            "FUNCTION MY_FUNC : INT
-            VAR_INPUT
-                x : INT;
-            END_VAR
-            VAR
-                buf : STRING[10];
-            END_VAR
-            MY_FUNC := 0;
-            END_FUNCTION",
-        );
-    }
-
-    #[test]
-    fn parse_when_function_var_constant_with_string_length_then_parses() {
-        let _lib = parse_text(
-            "FUNCTION MY_FUNC : INT
-            VAR CONSTANT
-                FILL : STRING[1] := '0';
-            END_VAR
-            MY_FUNC := 0;
-            END_FUNCTION",
-        );
-    }
-
-    #[test]
-    fn parse_when_function_var_in_out_with_string_length_then_parses() {
-        let _lib = parse_text(
-            "FUNCTION MY_FUNC : INT
-            VAR_IN_OUT
-                buf : STRING[255];
-            END_VAR
-            MY_FUNC := 0;
-            END_FUNCTION",
-        );
-    }
-
-    #[test]
-    fn parse_when_struct_member_with_string_length_then_parses() {
-        let _lib = parse_text(
-            "TYPE MY_STRUCT :
-            STRUCT
-                name : STRING[10];
-            END_STRUCT;
-            END_TYPE
-            FUNCTION MY_FUNC : INT
-            VAR_INPUT
-                x : INT;
-            END_VAR
-            MY_FUNC := 0;
-            END_FUNCTION",
-        );
-    }
-
-    #[test]
-    fn parse_when_task_with_interval_and_priority_then_builds_structure() {
-        let source = "
-        CONFIGURATION config
-            RESOURCE resource1 ON PLC
-                TASK my_task(INTERVAL := T#50ms, PRIORITY := 3);
-                PROGRAM instance1 WITH my_task : my_prg;
-            END_RESOURCE
-        END_CONFIGURATION";
-
-        let lib = parse_text(source);
-        let config = cast!(
-            &lib.elements[0],
-            LibraryElementKind::ConfigurationDeclaration
-        );
-        let task = &config.resource_decl[0].tasks[0];
-        assert_eq!(task.name, Id::from("my_task"));
-        assert_eq!(task.priority, 3);
-        assert!(task.interval.is_some());
-        assert!(task.single.is_none());
-    }
-
-    #[test]
-    fn parse_when_task_with_single_constant_then_builds_structure() {
-        let source = "
-        CONFIGURATION config
-            RESOURCE resource1 ON PLC
-                TASK event_task(SINGLE := 1, PRIORITY := 0);
-                PROGRAM instance1 WITH event_task : my_prg;
-            END_RESOURCE
-        END_CONFIGURATION";
-
-        let lib = parse_text(source);
-        let config = cast!(
-            &lib.elements[0],
-            LibraryElementKind::ConfigurationDeclaration
-        );
-        let task = &config.resource_decl[0].tasks[0];
-        assert_eq!(task.name, Id::from("event_task"));
-        assert_eq!(task.priority, 0);
-        assert!(task.interval.is_none());
-        assert!(task.single.is_some());
-        assert!(matches!(task.single, Some(DataSourceKind::Constant(_))));
-    }
-
-    #[test]
-    fn parse_when_task_with_single_and_interval_then_builds_structure() {
-        let source = "
-        CONFIGURATION config
-            RESOURCE resource1 ON PLC
-                TASK my_task(SINGLE := 1, INTERVAL := T#100ms, PRIORITY := 1);
-                PROGRAM instance1 WITH my_task : my_prg;
-            END_RESOURCE
-        END_CONFIGURATION";
-
-        let lib = parse_text(source);
-        let config = cast!(
-            &lib.elements[0],
-            LibraryElementKind::ConfigurationDeclaration
-        );
-        let task = &config.resource_decl[0].tasks[0];
-        assert_eq!(task.name, Id::from("my_task"));
-        assert_eq!(task.priority, 1);
-        assert!(task.interval.is_some());
-        assert!(task.single.is_some());
-    }
-
-    #[test]
-    fn parse_when_task_with_priority_only_then_builds_structure() {
-        let source = "
-        CONFIGURATION config
-            RESOURCE resource1 ON PLC
-                TASK free_task(PRIORITY := 5);
-                PROGRAM instance1 WITH free_task : my_prg;
-            END_RESOURCE
-        END_CONFIGURATION";
-
-        let lib = parse_text(source);
-        let config = cast!(
-            &lib.elements[0],
-            LibraryElementKind::ConfigurationDeclaration
-        );
-        let task = &config.resource_decl[0].tasks[0];
-        assert_eq!(task.name, Id::from("free_task"));
-        assert_eq!(task.priority, 5);
-        assert!(task.interval.is_none());
-        assert!(task.single.is_none());
-    }
-
-    #[test]
-    fn parse_when_task_with_interval_non_duration_then_parse_error() {
-        let source = "
-        CONFIGURATION config
-            RESOURCE resource1 ON PLC
-                TASK bad_task(INTERVAL := 42, PRIORITY := 1);
-                PROGRAM instance1 WITH bad_task : my_prg;
-            END_RESOURCE
-        END_CONFIGURATION";
-
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_mod_function_call_then_parses() {
-        parse_text(
-            "FUNCTION_BLOCK CALLER
-VAR
-    result : INT;
-    a : INT;
-    b : INT;
-END_VAR
-    result := MOD(a, b);
-END_FUNCTION_BLOCK",
-        );
-    }
-
-    #[test]
-    fn parse_when_and_function_call_then_parses() {
-        parse_text(
-            "FUNCTION_BLOCK CALLER
-VAR
-    result : BOOL;
-    a : BOOL;
-    b : BOOL;
-END_VAR
-    result := AND(a, b);
-END_FUNCTION_BLOCK",
-        );
-    }
-
-    #[test]
-    fn parse_when_or_function_call_then_parses() {
-        parse_text(
-            "FUNCTION_BLOCK CALLER
-VAR
-    result : BOOL;
-    a : BOOL;
-    b : BOOL;
-END_VAR
-    result := OR(a, b);
-END_FUNCTION_BLOCK",
-        );
-    }
-
-    #[test]
-    fn parse_when_xor_function_call_then_parses() {
-        parse_text(
-            "FUNCTION_BLOCK CALLER
-VAR
-    result : BOOL;
-    a : BOOL;
-    b : BOOL;
-END_VAR
-    result := XOR(a, b);
-END_FUNCTION_BLOCK",
-        );
-    }
-
-    #[test]
-    fn parse_when_bit_access_then_succeeds() {
-        let res = parse_resource("bit_access.st");
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    fn parse_when_bit_access_then_returns_bit_access_variable() {
-        let lib = parse_text(
-            "FUNCTION_BLOCK FB1
-VAR
-    x : WORD;
-    y : BOOL;
-END_VAR
-    y := x.0;
-END_FUNCTION_BLOCK",
-        );
-        let fb = cast!(
-            &lib.elements[0],
-            LibraryElementKind::FunctionBlockDeclaration
-        );
-        let s = cast!(&fb.body, FunctionBlockBodyKind::Statements);
-        let assign = cast!(&s.body[0], StmtKind::Assignment);
-        // The value (RHS) should be a bit access variable x.0
-        let var_sym = cast!(&assign.value.kind, ExprKind::Variable);
-        let var_kind = cast!(var_sym, Variable::Symbolic);
-        let var = cast!(var_kind, SymbolicVariableKind::BitAccess);
-        assert_eq!(var.variable.as_ref().to_string(), "x");
-        assert_eq!(var.index.value, 0);
-    }
-
-    #[test]
-    fn parse_when_struct_access_with_bit_access_support_then_still_succeeds() {
-        parse_text(
-            "FUNCTION_BLOCK FB1
-VAR
-    counter : CounterST;
-    result : INT;
-END_VAR
-    result := counter.OUT;
-END_FUNCTION_BLOCK",
-        );
-    }
-
-    /// Returns options with `allow_missing_semicolon` enabled. Used as a
-    /// function-pointer `#[case]` value for the parametrized dialect-flag tests.
-    fn with_missing_semicolon_flag() -> CompilerOptions {
-        CompilerOptions {
-            allow_missing_semicolon: true,
-            ..Default::default()
-        }
-    }
-
-    /// Returns options with `allow_empty_var_blocks` enabled.
-    fn with_empty_var_blocks_flag() -> CompilerOptions {
-        CompilerOptions {
-            allow_empty_var_blocks: true,
-            ..Default::default()
-        }
-    }
-
-    /// Dialect-flag acceptance tests.
-    ///
-    /// Each case: given a program that the default dialect rejects, parsing
-    /// should succeed when the corresponding `--allow-*` flag is enabled.
-    /// Collapses 14 hand-written tests (cargo-dupes group `c3c13bfc`) into
-    /// one table; each row is still run as an individually-named test.
-    #[rstest]
-    // --allow-missing-semicolon: trailing STRUCT/END_IF/END_WHILE/etc. without `;`
-    #[case::struct_without_trailing_semicolon(
-        "TYPE MY_POINT :
-STRUCT
-    X : REAL;
-    Y : REAL;
-END_STRUCT
-END_TYPE
-
-PROGRAM main
-VAR
-    pt : MY_POINT;
-END_VAR
-    pt.X := 1.0;
-    pt.Y := 2.0;
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::end_if_without_semicolon(
-        "PROGRAM main
-VAR
-    x : BOOL;
-END_VAR
-    IF x THEN
-        x := FALSE;
-    END_IF
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::end_while_without_semicolon(
-        "PROGRAM main
-VAR
-    x : BOOL;
-END_VAR
-    WHILE x DO
-        x := FALSE;
-    END_WHILE
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::end_for_without_semicolon(
-        "PROGRAM main
-VAR
-    i : INT;
-END_VAR
-    FOR i := 0 TO 10 DO
-        i := i;
-    END_FOR
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::end_case_without_semicolon(
-        "PROGRAM main
-VAR
-    x : INT;
-END_VAR
-    CASE x OF
-        1: x := 2;
-    END_CASE
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::end_repeat_without_semicolon(
-        "PROGRAM main
-VAR
-    x : BOOL;
-END_VAR
-    REPEAT
-        x := FALSE;
-    UNTIL x
-    END_REPEAT
-END_PROGRAM",
-        with_missing_semicolon_flag
-    )]
-    #[case::function_end_if_without_semicolon(
-        "FUNCTION MY_FUNC : REAL
-VAR_INPUT
-    x : INT;
-END_VAR
-IF x > 0 THEN
-    MY_FUNC := 1.0;
-ELSE
-    MY_FUNC := 0.0;
-END_IF
-END_FUNCTION",
-        with_missing_semicolon_flag
-    )]
-    // --allow-empty-var-blocks: empty VAR / VAR_INPUT / VAR_OUTPUT / etc.
-    #[case::empty_var_in_function(
-        "
-FUNCTION myFunc : INT
-VAR
-END_VAR
-    myFunc := 1;
-END_FUNCTION",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_input(
-        "
-FUNCTION myFunc : INT
-VAR_INPUT
-END_VAR
-    myFunc := 1;
-END_FUNCTION",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_output(
-        "
-FUNCTION myFunc : INT
-VAR_OUTPUT
-END_VAR
-    myFunc := 1;
-END_FUNCTION",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_in_out(
-        "
-FUNCTION myFunc : INT
-VAR_IN_OUT
-END_VAR
-    myFunc := 1;
-END_FUNCTION",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_constant(
-        "
-FUNCTION myFunc : INT
-VAR CONSTANT
-END_VAR
-    myFunc := 1;
-END_FUNCTION",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_in_program(
-        "
-PROGRAM main
-VAR
-END_VAR
-END_PROGRAM",
-        with_empty_var_blocks_flag
-    )]
-    #[case::empty_var_in_function_block(
-        "
-FUNCTION_BLOCK myFb
-VAR
-END_VAR
-END_FUNCTION_BLOCK",
-        with_empty_var_blocks_flag
-    )]
-    fn parse_with_dialect_flag_then_ok(
-        #[case] source: &str,
-        #[case] build_options: fn() -> CompilerOptions,
-    ) {
-        let options = build_options();
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn parse_when_struct_without_trailing_semicolon_and_flag_disabled_then_err() {
-        let source = "TYPE MY_POINT :
-STRUCT
-    X : REAL;
-    Y : REAL;
-END_STRUCT
-END_TYPE
-
-PROGRAM main
-VAR
-    pt : MY_POINT;
-END_VAR
-    pt.X := 1.0;
-    pt.Y := 2.0;
-END_PROGRAM";
-
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    fn parse_text_edition3(source: &str) -> Library {
-        let options = CompilerOptions {
-            allow_iec_61131_3_2013: true,
-            ..CompilerOptions::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-        result.unwrap()
-    }
-
-    fn parse_text_reference_to(source: &str) -> Library {
-        let options = CompilerOptions {
-            allow_reference_to: true,
-            ..CompilerOptions::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-        result.unwrap()
-    }
-
-    /// The single (non-FB-call) statement in a program body.
-    fn only_statement(lib: &Library) -> &StmtKind {
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let stmts = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(stmts.body.len(), 1, "expected exactly one statement");
-        &stmts.body[0]
-    }
-
-    // --- REF= disambiguation against other uses of `=` -----------------------
-
-    /// A variable *named* `REF` assigned with `:=` is a normal assignment, not a
-    /// `REF=` binding — even with `--allow-reference-to` active.
-    #[test]
-    fn ref_bind_when_variable_named_ref_assigned_then_normal_assignment() {
-        let lib = parse_text_reference_to(
-            "PROGRAM main
-VAR
-    REF : INT;
-END_VAR
-    REF := 5;
-END_PROGRAM",
-        );
-        let assignment = cast!(only_statement(&lib), StmtKind::Assignment);
-        assert!(
-            !assignment.ref_bind,
-            "must not be treated as a REF= binding"
-        );
-        assert_eq!(assignment.target.to_string(), "REF");
-    }
-
-    /// An `=` equality comparison in a condition still parses (the REF= rule
-    /// only fires on `REF` immediately followed by `=`).
-    #[test]
-    fn ref_bind_when_equality_in_condition_then_parses_as_comparison() {
-        let lib = parse_text_reference_to(
-            "PROGRAM main
-VAR
-    a : INT;
-    b : INT;
-    c : INT;
-END_VAR
-    IF a = b THEN
-        c := 1;
-    END_IF;
-END_PROGRAM",
-        );
-        // The single top-level statement is the IF, not an assignment.
-        cast!(only_statement(&lib), StmtKind::If);
-    }
-
-    /// An `=` comparison on the right-hand side of `:=` still parses as a normal
-    /// assignment whose value is the comparison.
-    #[test]
-    fn ref_bind_when_equality_in_rhs_then_normal_assignment() {
-        let lib = parse_text_reference_to(
-            "PROGRAM main
-VAR
-    a : INT;
-    c : INT;
-    result : BOOL;
-END_VAR
-    result := a = c;
-END_PROGRAM",
-        );
-        let assignment = cast!(only_statement(&lib), StmtKind::Assignment);
-        assert!(!assignment.ref_bind);
-    }
-
-    /// `REF=` binds when `REF` is the keyword token (both `--allow-ref-to` and
-    /// `--allow-reference-to` active, e.g. the `codesys` dialect).
-    #[test]
-    fn ref_bind_when_ref_is_keyword_token_then_binds() {
-        let options = CompilerOptions {
-            allow_ref_to: true,
-            allow_reference_to: true,
-            ..CompilerOptions::default()
-        };
-        let source = "PROGRAM main
-VAR
-    x : INT;
-    r : REFERENCE TO INT;
-END_VAR
-    r REF= x;
-END_PROGRAM";
-        let lib = parse_program(source, &FileId::default(), &options).unwrap();
-        let assignment = cast!(only_statement(&lib), StmtKind::Assignment);
-        assert!(assignment.ref_bind);
-        let referent = cast!(&assignment.value.kind, ExprKind::Ref);
-        assert_eq!(referent.to_string(), "x");
-    }
-
-    /// The `REF=` operator is case-insensitive (`ref=` binds too).
-    #[test]
-    fn ref_bind_when_lowercase_operator_then_binds() {
-        let lib = parse_text_reference_to(
-            "PROGRAM main
-VAR
-    x : INT;
-    r : REFERENCE TO INT;
-END_VAR
-    r ref= x;
-END_PROGRAM",
-        );
-        let assignment = cast!(only_statement(&lib), StmtKind::Assignment);
-        assert!(assignment.ref_bind);
-    }
-
-    /// A space between `REF` and `=` is not the binding operator: `REF` and `=`
-    /// must be adjacent, so `x REF = y` is rejected.
-    #[test]
-    fn ref_bind_when_space_between_ref_and_equals_then_error() {
-        let options = CompilerOptions {
-            allow_reference_to: true,
-            ..CompilerOptions::default()
-        };
-        let source = "PROGRAM main
-VAR
-    x : REFERENCE TO INT;
-    y : INT;
-END_VAR
-    x REF = y;
-END_PROGRAM";
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_err(), "REF = with a space must be rejected");
-    }
-
-    /// The right-hand side of `REF=` must be a variable (address-of), matching
-    /// `REF()` semantics — a literal is rejected rather than silently accepted.
-    #[test]
-    fn ref_bind_when_rhs_is_literal_then_error() {
-        let options = CompilerOptions {
-            allow_reference_to: true,
-            ..CompilerOptions::default()
-        };
-        let source = "PROGRAM main
-VAR
-    r : REFERENCE TO INT;
-END_VAR
-    r REF= 5;
-END_PROGRAM";
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_err(), "REF= to a literal must be rejected");
-    }
-
-    /// A referent *named* `REF` can be bound: `x REF= REF` binds `x` to the
-    /// variable `REF`.
-    #[test]
-    fn ref_bind_when_referent_named_ref_then_binds() {
-        let lib = parse_text_reference_to(
-            "PROGRAM main
-VAR
-    REF : INT;
-    x : REFERENCE TO INT;
-END_VAR
-    x REF= REF;
-END_PROGRAM",
-        );
-        let assignment = cast!(only_statement(&lib), StmtKind::Assignment);
-        assert!(assignment.ref_bind);
-        let referent = cast!(&assignment.value.kind, ExprKind::Ref);
-        assert_eq!(referent.to_string(), "REF");
-    }
-
-    #[test]
-    fn parse_when_ref_to_int_type_decl_then_ok() {
-        let lib = parse_text_edition3("TYPE IntRef : REF_TO INT; END_TYPE");
-        assert_eq!(lib.elements.len(), 1);
-        let dt = cast!(&lib.elements[0], LibraryElementKind::DataTypeDeclaration);
-        let decl = cast!(dt, DataTypeDeclarationKind::Reference);
-        assert_eq!(decl.type_name.to_string(), "IntRef");
-        assert_eq!(decl.target.type_name().unwrap().to_string(), "INT");
-    }
-
-    #[test]
-    fn parse_when_ref_to_var_decl_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    x : REF_TO INT;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-        assert!(matches!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Reference(_)
-        ));
-    }
-
-    #[test]
-    fn parse_when_ref_to_var_decl_with_null_init_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    x : REF_TO INT := NULL;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let ref_init = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Reference
-        );
-        assert!(matches!(
-            ref_init.initial_value,
-            Some(dsl::common::ReferenceInitialValue::Null(_))
-        ));
-    }
-
-    #[test]
-    fn parse_when_ref_to_var_decl_with_ref_init_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    counter : INT;
-    x : REF_TO INT := REF(counter);
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let ref_init = cast!(
-            &prog.variables[1].initializer,
-            InitialValueAssignmentKind::Reference
-        );
-        assert!(matches!(
-            ref_init.initial_value,
-            Some(dsl::common::ReferenceInitialValue::Ref(_))
-        ));
-    }
-
-    #[test]
-    fn parse_when_ref_to_array_var_decl_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    x : REF_TO ARRAY[1..10] OF INT;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-        let ref_init = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Reference
-        );
-        assert!(matches!(ref_init.target, ReferenceTarget::Array(_)));
-    }
-
-    #[test]
-    fn parse_when_ref_to_array_type_decl_then_ok() {
-        let lib = parse_text_edition3("TYPE ArrRef : REF_TO ARRAY[0..3] OF BYTE; END_TYPE");
-        assert_eq!(lib.elements.len(), 1);
-        let dt = cast!(&lib.elements[0], LibraryElementKind::DataTypeDeclaration);
-        let decl = cast!(dt, DataTypeDeclarationKind::Reference);
-        assert_eq!(decl.type_name.to_string(), "ArrRef");
-        assert!(matches!(decl.target, ReferenceTarget::Array(_)));
-    }
-
-    #[test]
-    fn parse_when_ref_operator_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    counter : INT;
-    x : REF_TO INT;
-END_VAR
-    x := REF(counter);
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert!(!s.body.is_empty());
-    }
-
-    #[test]
-    fn parse_when_deref_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    myRef : REF_TO INT;
-    value : INT;
-END_VAR
-    value := myRef^;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_deref_assign_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    myRef : REF_TO INT;
-END_VAR
-    myRef^ := 42;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-        let assignment = cast!(&s.body[0], StmtKind::Assignment);
-        assert!(assignment.deref);
-    }
-
-    #[test]
-    fn parse_when_null_literal_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    myRef : REF_TO INT;
-END_VAR
-    myRef := NULL;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_null_comparison_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    myRef : REF_TO INT;
-    x : INT;
-END_VAR
-    IF myRef <> NULL THEN
-        x := 1;
-    END_IF;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_xor_still_works_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    a : BOOL;
-    b : BOOL;
-    result : BOOL;
-END_VAR
-    result := a XOR b;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_deref_then_xor_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    myRef : REF_TO BOOL;
-    b : BOOL;
-    result : BOOL;
-END_VAR
-    result := myRef^ XOR b;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let s = cast!(&prog.body, FunctionBlockBodyKind::Statements);
-        assert_eq!(s.body.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_deref_array_subscript_then_ok() {
-        let lib = parse_text_edition3(
-            "FUNCTION my_func : INT
-VAR_INPUT
-    PT : REF_TO ARRAY[0..10] OF BYTE;
-END_VAR
-    my_func := 0;
-    PT^[0] := BYTE#0;
-END_FUNCTION
-
-PROGRAM test_main
-VAR
-    result : INT;
-END_VAR
-    result := 0;
-END_PROGRAM",
-        );
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let stmts = &func.body;
-        assert_eq!(stmts.len(), 2);
-        let assignment = cast!(&stmts[1], StmtKind::Assignment);
-        // Target should be Array(Deref(Named(PT)), [0])
-        let sym = cast!(&assignment.target, Variable::Symbolic);
-        let arr = cast!(sym, SymbolicVariableKind::Array);
-        let d = cast!(
-            arr.subscripted_variable.as_ref(),
-            SymbolicVariableKind::Deref
-        );
-        let n = cast!(d.variable.as_ref(), SymbolicVariableKind::Named);
-        assert_eq!(n.name.to_string(), "PT");
-    }
-
-    #[test]
-    fn parse_when_deref_array_subscript_in_expression_then_ok() {
-        let lib = parse_text_edition3(
-            "FUNCTION my_func : BYTE
-VAR_INPUT
-    PT : REF_TO ARRAY[0..10] OF BYTE;
-END_VAR
-    my_func := PT^[0];
-END_FUNCTION
-
-PROGRAM test_main
-VAR
-    result : BYTE;
-END_VAR
-    result := BYTE#0;
-END_PROGRAM",
-        );
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        let stmts = &func.body;
-        assert_eq!(stmts.len(), 1);
-        let assignment = cast!(&stmts[0], StmtKind::Assignment);
-        // Value should be Variable(Array(Deref(Named(PT)), [0]))
-        let var = cast!(&assignment.value.kind, ExprKind::Variable);
-        let sym = cast!(var, Variable::Symbolic);
-        let arr = cast!(sym, SymbolicVariableKind::Array);
-        let d = cast!(
-            arr.subscripted_variable.as_ref(),
-            SymbolicVariableKind::Deref
-        );
-        let n = cast!(d.variable.as_ref(), SymbolicVariableKind::Named);
-        assert_eq!(n.name.to_string(), "PT");
-    }
-
-    #[test]
-    fn parse_when_ldate_literal_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    d : LDATE;
-END_VAR
-    d := LDATE#2024-01-20;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_ltod_literal_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    t : LTOD;
-END_VAR
-    t := LTOD#14:30:20;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_ldt_literal_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    my_dt : LDT;
-END_VAR
-    my_dt := LDT#2024-01-20-15:30:22;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_ltime_of_day_long_form_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    t : LTIME_OF_DAY;
-END_VAR
-    t := LTOD#10:00:00;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_ldate_and_time_long_form_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    my_dt : LDATE_AND_TIME;
-END_VAR
-    my_dt := LDT#2024-01-20-15:30:22;
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_array_of_ref_to_then_ok() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    data : ARRAY[0..3] OF REF_TO BYTE;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 1);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert!(subranges.ref_to.is_some());
-        assert_eq!(subranges.type_name.to_type_name().to_string(), "BYTE");
-        assert_eq!(subranges.ranges.len(), 1);
-    }
-
-    #[test]
-    fn parse_when_array_of_ref_to_type_decl_then_ok() {
-        let lib = parse_text_edition3("TYPE MyArr : ARRAY[1..5] OF REF_TO INT; END_TYPE");
-        let dt = cast!(&lib.elements[0], LibraryElementKind::DataTypeDeclaration);
-        let arr = cast!(dt, DataTypeDeclarationKind::Array);
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert!(subranges.ref_to.is_some());
-        assert_eq!(subranges.type_name.to_type_name().to_string(), "INT");
-    }
-
-    #[test]
-    fn parse_when_array_without_ref_to_then_ref_to_is_false() {
-        let lib = parse_text_edition3(
-            "PROGRAM main
-VAR
-    data : ARRAY[0..3] OF BYTE;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert!(subranges.ref_to.is_none());
-    }
-
-    #[test]
-    fn parse_when_array_of_string_with_size_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR
-    names : ARRAY[1..3] OF STRING[10];
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert_eq!(subranges.type_name.to_type_name().to_string(), "STRING");
-        assert_eq!(subranges.ranges.len(), 1);
-        let spec = cast!(&subranges.type_name, ArrayElementType::String);
-        assert_eq!(
-            spec.length.as_ref().unwrap().as_integer().unwrap().value,
-            10
-        );
-    }
-
-    #[test]
-    fn parse_when_array_of_wstring_with_size_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR
-    names : ARRAY[1..3] OF WSTRING[20];
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert_eq!(subranges.type_name.to_type_name().to_string(), "WSTRING");
-        let spec = cast!(&subranges.type_name, ArrayElementType::WString);
-        assert_eq!(
-            spec.length.as_ref().unwrap().as_integer().unwrap().value,
-            20
-        );
-    }
-
-    #[test]
-    fn parse_when_array_of_string_without_size_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR
-    names : ARRAY[1..3] OF STRING;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        let spec = cast!(&subranges.type_name, ArrayElementType::String);
-        assert!(spec.length.is_none());
-    }
-
-    #[test]
-    fn parse_when_multidim_array_of_string_with_size_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR
-    weekdays : ARRAY[1..3, 1..7] OF STRING[10];
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        let arr = cast!(
-            &prog.variables[0].initializer,
-            InitialValueAssignmentKind::Array
-        );
-        let subranges = cast!(&arr.spec, SpecificationKind::Inline);
-        assert_eq!(subranges.ranges.len(), 2);
-        let spec = cast!(&subranges.type_name, ArrayElementType::String);
-        assert_eq!(
-            spec.length.as_ref().unwrap().as_integer().unwrap().value,
-            10
-        );
-    }
-
-    #[test]
-    fn parse_when_empty_var_in_function_without_flag_then_error() {
-        let source = "
-FUNCTION myFunc : INT
-VAR
-END_VAR
-    myFunc := 1;
-END_FUNCTION";
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_time_function_call_and_flag_enabled_then_ok() {
-        let source = "FUNCTION MY_FUNC : DWORD
-VAR
-    tx : TIME;
-END_VAR
-tx := TIME();
-MY_FUNC := TIME_TO_DWORD(tx);
-END_FUNCTION";
-
-        let options = CompilerOptions {
-            allow_time_as_function_name: true,
-            ..Default::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_when_time_function_call_and_flag_disabled_then_err() {
-        let source = "FUNCTION MY_FUNC : DWORD
-VAR
-    tx : TIME;
-END_VAR
-tx := TIME();
-END_FUNCTION";
-
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_time_duration_literal_and_flag_enabled_then_ok() {
-        let source = "PROGRAM main
-VAR
-    t : TIME;
-END_VAR
-    t := TIME#5s;
-END_PROGRAM";
-
-        let options = CompilerOptions {
-            allow_time_as_function_name: true,
-            ..Default::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_when_time_type_decl_and_flag_enabled_then_ok() {
-        let source = "PROGRAM main
-VAR
-    tx : TIME;
-END_VAR
-    tx := TIME#1s;
-END_PROGRAM";
-
-        let options = CompilerOptions {
-            allow_time_as_function_name: true,
-            ..Default::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_when_time_function_declaration_and_flag_enabled_then_ok() {
-        let source = "FUNCTION TIME : TIME
-VAR
-    t : TIME;
-END_VAR
-TIME := T#0s;
-END_FUNCTION
-
-PROGRAM main
-VAR
-    t : TIME;
-END_VAR
-t := TIME();
-END_PROGRAM";
-
-        let options = CompilerOptions {
-            allow_time_as_function_name: true,
-            ..Default::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "Parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_when_time_function_declaration_and_flag_disabled_then_err() {
-        let source = "FUNCTION TIME : TIME
-TIME := T#0s;
-END_FUNCTION";
-
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_function_with_var_temp_then_succeeds() {
-        let lib = parse_text(
-            "FUNCTION my_func : DINT
-VAR_INPUT
-    a : DINT;
-END_VAR
-VAR_TEMP
-    temp : DINT;
-END_VAR
-    temp := a * 2;
-    my_func := temp;
-END_FUNCTION",
-        );
-        let func = cast!(&lib.elements[0], LibraryElementKind::FunctionDeclaration);
-        assert_eq!(func.variables.len(), 2);
-        assert_eq!(func.variables[0].var_type, VariableType::Input);
-        assert_eq!(func.variables[1].var_type, VariableType::VarTemp);
-    }
-
-    #[test]
-    fn parse_when_function_block_with_var_temp_then_succeeds() {
-        let lib = parse_text(
-            "FUNCTION_BLOCK my_fb
-VAR_TEMP
-    t : INT;
-END_VAR
-    t := 42;
-END_FUNCTION_BLOCK",
-        );
-        let fb = cast!(
-            &lib.elements[0],
-            LibraryElementKind::FunctionBlockDeclaration
-        );
-        assert_eq!(fb.variables.len(), 1);
-        assert_eq!(fb.variables[0].var_type, VariableType::VarTemp);
-    }
-
-    #[test]
-    fn parse_when_program_with_var_temp_then_fails() {
-        let source = "PROGRAM main
-VAR_TEMP
-    t : INT;
-END_VAR
-    t := 42;
-END_PROGRAM";
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_program_mixed_located_and_non_located_vars_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR
-    Motor : BOOL;
-    xStart AT %IX0.0 : BOOL;
-    xStop AT %IX0.1 : BOOL;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 3);
-        assert!(matches!(
-            &prog.variables[0].identifier,
-            VariableIdentifier::Symbol(_)
-        ));
-        assert!(matches!(
-            &prog.variables[1].identifier,
-            VariableIdentifier::Direct(_)
-        ));
-        assert!(matches!(
-            &prog.variables[2].identifier,
-            VariableIdentifier::Direct(_)
-        ));
-    }
-
-    #[test]
-    fn parse_when_program_mixed_vars_with_retain_qualifier_then_ok() {
-        let lib = parse_text(
-            "PROGRAM main
-VAR RETAIN
-    counter : INT;
-    saved AT %MW0 : INT;
-END_VAR
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 2);
-        assert_eq!(prog.variables[0].qualifier, DeclarationQualifier::Retain);
-        assert_eq!(prog.variables[1].qualifier, DeclarationQualifier::Retain);
-    }
-
-    #[test]
-    fn parse_when_program_motor_control_style_then_ok() {
-        let lib = parse_text(
-            "TYPE
-  MotorState : (STOPPED, RUNNING, FAULTED);
-END_TYPE
-
-FUNCTION_BLOCK FB_MotorControl
-  VAR_INPUT
-    START_PB : BOOL;
-    STOP_PB : BOOL;
-    OL_CONTACT : BOOL;
-    FAULT_RESET : BOOL;
-  END_VAR
-  VAR_OUTPUT
-    CONTACTOR : BOOL;
-    RUN_LAMP : BOOL;
-    FAULT_LAMP : BOOL;
-  END_VAR
-  VAR
-    Seal : BOOL;
-  END_VAR
-
-  IF NOT OL_CONTACT THEN
-    Seal := FALSE;
-  ELSE
-    IF START_PB AND STOP_PB THEN
-      Seal := TRUE;
-    END_IF;
-    IF NOT STOP_PB THEN
-      Seal := FALSE;
-    END_IF;
-  END_IF;
-
-  CONTACTOR := Seal;
-  RUN_LAMP := CONTACTOR;
-  FAULT_LAMP := NOT OL_CONTACT;
-END_FUNCTION_BLOCK
-
-PROGRAM PLC_PRG
-  VAR
-    Motor : FB_MotorControl;
-    xStart AT %IX0.0 : BOOL;
-    xStop AT %IX0.1 : BOOL;
-    xOverload AT %IX0.2 : BOOL;
-    xReset AT %IX0.3 : BOOL;
-    yContactor AT %QX0.0 : BOOL;
-    yRunLamp AT %QX0.1 : BOOL;
-    yFaultLamp AT %QX0.2 : BOOL;
-  END_VAR
-
-  Motor(
-    START_PB := xStart,
-    STOP_PB := xStop,
-    OL_CONTACT := xOverload,
-    FAULT_RESET := xReset,
-    CONTACTOR => yContactor,
-    RUN_LAMP => yRunLamp,
-    FAULT_LAMP => yFaultLamp
-  );
-END_PROGRAM",
-        );
-        let prog = cast!(&lib.elements[2], LibraryElementKind::ProgramDeclaration);
-        assert_eq!(prog.variables.len(), 8);
-        assert!(matches!(
-            &prog.variables[0].identifier,
-            VariableIdentifier::Symbol(_)
-        ));
-        assert!(matches!(
-            &prog.variables[1].identifier,
-            VariableIdentifier::Direct(_)
-        ));
-    }
-
-    // ---------------------------------------------------------------------
-    // REQ-PAB: IEC 61131-3:2013 partial-access bit syntax (.%Xn).
-    // See specs/design/partial-access-bit-syntax.md.
-    // ---------------------------------------------------------------------
-
-    fn opts_with_partial_access() -> CompilerOptions {
-        CompilerOptions {
-            allow_partial_access_syntax: true,
-            ..CompilerOptions::default()
-        }
-    }
-
-    fn wrap_program(body: &str) -> String {
-        format!(
-            "PROGRAM main\nVAR\n  b : BYTE;\n  r : BOOL;\n  arr : ARRAY[0..1] OF BYTE;\n  s : MY_STRUCT;\nEND_VAR\n{}\nEND_PROGRAM",
-            body
-        )
-    }
-
-    /// REQ-PAB-001: The lexer tokenizes `%X<digits>` as `PartialAccessBit`.
-    #[test]
-    fn lexer_spec_req_pab_001_percent_x_digits_tokenizes_as_partial_access_bit() {
-        use crate::token::TokenType;
-        let (tokens, _) = crate::tokenize_program(
-            "PROGRAM p VAR b : BYTE; END_VAR b.%X3; END_PROGRAM",
-            &FileId::default(),
-            &opts_with_partial_access(),
-            0,
-            0,
-        );
-        let uppercase_hit = tokens
-            .iter()
-            .any(|t| t.token_type == TokenType::PartialAccessBit && t.text == "%X3");
-        assert!(uppercase_hit, "tokens = {:?}", tokens);
-
-        // Case insensitivity.
-        let (tokens_lower, _) = crate::tokenize_program(
-            "PROGRAM p VAR b : BYTE; END_VAR b.%x3; END_PROGRAM",
-            &FileId::default(),
-            &opts_with_partial_access(),
-            0,
-            0,
-        );
-        assert!(tokens_lower
-            .iter()
-            .any(|t| t.token_type == TokenType::PartialAccessBit && t.text == "%x3"));
-    }
-
-    /// REQ-PAB-002: `%X0` does not cannibalize `DirectAddress` tokens like `%IX0.0`.
-    #[test]
-    fn lexer_spec_req_pab_002_direct_address_still_takes_precedence() {
-        use crate::token::TokenType;
-        let (tokens, _) = crate::tokenize_program(
-            "PROGRAM p VAR x AT %IX0.0 : BOOL; END_VAR END_PROGRAM",
-            &FileId::default(),
-            &opts_with_partial_access(),
-            0,
-            0,
-        );
-        assert!(tokens
-            .iter()
-            .any(|t| t.token_type == TokenType::DirectAddress && t.text == "%IX0.0"));
-        assert!(!tokens
-            .iter()
-            .any(|t| t.token_type == TokenType::PartialAccessBit));
-    }
-
-    /// REQ-PAB-010: `.%Xn` is accepted on a simple variable.
-    #[test]
-    fn parser_spec_req_pab_010_dot_percent_x_accepted_on_simple_var() {
-        let src = wrap_program("r := b.%X0;");
-        let result = parse_program(&src, &FileId::default(), &opts_with_partial_access());
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    /// REQ-PAB-011: `.%Xn` is accepted after an array subscript — the user's case.
-    #[test]
-    fn parser_spec_req_pab_011_dot_percent_x_accepted_after_array_subscript() {
-        let src = wrap_program("r := arr[0].%X0;");
-        let result = parse_program(&src, &FileId::default(), &opts_with_partial_access());
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    /// REQ-PAB-012: `.%Xn` is accepted after a struct field access.
-    #[test]
-    fn parser_spec_req_pab_012_dot_percent_x_accepted_after_struct_field() {
-        // Define MY_STRUCT with a BYTE field so the program type-checks later,
-        // though this test only exercises the parser surface.
-        let src = "
-TYPE MY_STRUCT : STRUCT f : BYTE; END_STRUCT; END_TYPE
-PROGRAM main
-VAR
-  s : MY_STRUCT;
-  r : BOOL;
-END_VAR
-  r := s.f.%X0;
-END_PROGRAM
-";
-        let result = parse_program(src, &FileId::default(), &opts_with_partial_access());
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    /// REQ-PAB-020: `x.%Xn` and `x.n` produce equal AST subtrees.
-    #[test]
-    fn parser_spec_req_pab_020_dot_percent_x_and_dot_n_produce_equal_ast() {
-        let long = wrap_program("r := b.%X3;");
-        let short = wrap_program("r := b.3;");
-        let long_lib =
-            parse_program(&long, &FileId::default(), &opts_with_partial_access()).unwrap();
-        let short_lib =
-            parse_program(&short, &FileId::default(), &CompilerOptions::default()).unwrap();
-
-        // Dig out the single assignment's RHS symbolic variable from each AST
-        // and compare structurally.
-        let long_prog = cast!(
-            &long_lib.elements[0],
-            LibraryElementKind::ProgramDeclaration
-        );
-        let short_prog = cast!(
-            &short_lib.elements[0],
-            LibraryElementKind::ProgramDeclaration
-        );
-        assert_eq!(long_prog.body, short_prog.body);
-    }
-
-    /// REQ-PAB-050: When the flag is off, `.%Xn` produces
-    /// `PartialAccessSyntaxDisabled` (P4033) — not a lexer-level P0003.
-    #[test]
-    fn parser_spec_req_pab_050_disabled_flag_produces_partial_access_syntax_disabled() {
-        let src = wrap_program("r := b.%X0;");
-        let result = parse_program(&src, &FileId::default(), &CompilerOptions::default());
-        match result {
-            Ok(_) => panic!("expected error, got Ok"),
-            Err(d) => {
-                assert_eq!(
-                    d.code,
-                    "P4033",
-                    "expected P4033 PartialAccessSyntaxDisabled, got {}: {}",
-                    d.code,
-                    d.description(),
-                );
-            }
-        }
-    }
-
-    // Duration literal conformance tests — see specs/design/time-literals.md.
-
-    fn duration_program(literal: &str) -> String {
-        format!(
-            "FUNCTION fun:TIME\nVAR\n    tv : TIME := {literal};\nEND_VAR\nfun := tv;\nEND_FUNCTION"
-        )
-    }
-
-    fn extract_duration(library: &Library) -> &DurationLiteral {
-        let func = cast!(
-            &library.elements[0],
-            LibraryElementKind::FunctionDeclaration
-        );
-        let initializer = &func.variables[0].initializer;
-        let simple = cast!(initializer, InitialValueAssignmentKind::Simple);
-        let constant = simple.initial_value.as_ref().expect("initializer");
-        cast!(constant, ConstantKind::Duration)
-    }
-
-    #[rstest]
-    #[case::lower_t_lower("t#5s")]
-    #[case::upper_t_lower("T#5s")]
-    #[case::keyword_upper("TIME#5s")]
-    #[case::keyword_lower("time#5s")]
-    #[case::keyword_mixed("Time#5s")]
-    fn duration_spec_req_tl_002_prefix_case_insensitive(#[case] literal: &str) {
-        // REQ-TL-002: prefix is recognized case-insensitively.
-        let source = duration_program(literal);
-        let result = parse_program(&source, &FileId::default(), &CompilerOptions::default());
-        assert!(
-            result.is_ok(),
-            "parse failed for {literal}: {:?}",
-            result.err()
-        );
-    }
-
-    #[rstest]
-    #[case("T#5us")]
-    #[case("T#5ns")]
-    fn duration_spec_req_tl_010_unsupported_unit_rejected(#[case] literal: &str) {
-        // REQ-TL-010: units outside {d, h, m, s, ms} are parse errors.
-        let source = duration_program(literal);
-        let result = parse_program(&source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err(), "expected parse error for {literal}");
-    }
-
-    #[rstest]
-    #[case::s("T#5S", Duration::seconds(5))]
-    #[case::ms("T#100MS", Duration::milliseconds(100))]
-    #[case::h("T#1H", Duration::hours(1))]
-    #[case::d("T#1D", Duration::days(1))]
-    #[case::m("T#30M", Duration::minutes(30))]
-    fn duration_spec_req_tl_011_unit_suffix_uppercase_accepted(
-        #[case] literal: &str,
-        #[case] expected: Duration,
-    ) {
-        // REQ-TL-011: unit suffixes are case-insensitive.
-        let source = duration_program(literal);
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        assert_eq!(extract_duration(&library).interval, expected);
-    }
-
-    #[rstest]
-    #[case::ms_capital_m("T#500Ms", Duration::milliseconds(500))]
-    #[case::ms_capital_s("T#500mS", Duration::milliseconds(500))]
-    fn duration_spec_req_tl_011_unit_suffix_mixed_case_accepted(
-        #[case] literal: &str,
-        #[case] expected: Duration,
-    ) {
-        // REQ-TL-011: mixed-case unit suffixes parse identically to lowercase.
-        let source = duration_program(literal);
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        assert_eq!(extract_duration(&library).interval, expected);
-    }
-
-    #[test]
-    fn duration_spec_req_tl_012_ms_matched_before_m() {
-        // REQ-TL-012: `T#100ms` is 100 milliseconds, not 100 minutes.
-        let source = duration_program("T#100ms");
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        assert_eq!(
-            extract_duration(&library).interval,
-            Duration::milliseconds(100)
-        );
-    }
-
-    #[test]
-    #[ignore = "compound interval grammar is unreachable; see specs/design/time-literals.md Future Work"]
-    fn duration_spec_req_tl_021_compound_interval() {
-        // REQ-TL-021: compound interval with parts in descending magnitude.
-        let source = duration_program("T#1d2h30m15s500ms");
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        let expected = Duration::days(1)
-            + Duration::hours(2)
-            + Duration::minutes(30)
-            + Duration::seconds(15)
-            + Duration::milliseconds(500);
-        assert_eq!(extract_duration(&library).interval, expected);
-    }
-
-    #[rstest]
-    #[case::lower("T#1d_2h_30m_5s_100ms")]
-    #[case::upper("T#1D_2H_30M_5S_100MS")]
-    #[ignore = "compound interval grammar is unreachable; see specs/design/time-literals.md Future Work"]
-    fn duration_spec_req_tl_022_compound_with_underscore(#[case] literal: &str) {
-        // REQ-TL-022: optional `_` separator in compound intervals.
-        let source = duration_program(literal);
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        let expected = Duration::days(1)
-            + Duration::hours(2)
-            + Duration::minutes(30)
-            + Duration::seconds(5)
-            + Duration::milliseconds(100);
-        assert_eq!(extract_duration(&library).interval, expected);
-    }
-
-    #[test]
-    fn duration_spec_req_tl_023_negative_duration() {
-        // REQ-TL-023: optional leading `-` negates the interval.
-        let source = duration_program("T#-5s");
-        let library =
-            parse_program(&source, &FileId::default(), &CompilerOptions::default()).unwrap();
-        assert_eq!(extract_duration(&library).interval, Duration::seconds(-5));
-    }
-
-    #[rstest]
-    #[case("D#2026-01-01")]
-    #[case("d#2026-01-01")]
-    fn date_prefix_case_insensitive(#[case] literal: &str) {
-        // Sanity check for the `dt_sep("D")` simplification.
-        let source = format!(
-            "FUNCTION fun:DATE\nVAR\n    dv : DATE := {literal};\nEND_VAR\nfun := dv;\nEND_FUNCTION"
-        );
-        let result = parse_program(&source, &FileId::default(), &CompilerOptions::default());
-        assert!(
-            result.is_ok(),
-            "parse failed for {literal}: {:?}",
-            result.err()
-        );
-    }
-
-    // ---------------------------------------------------------------------
-    // TwinCAT/Siemens `{ ... }` pragma skipping.
-    // See specs/plans/2026-07-18-twincat-pragma-skipping.md.
-    // ---------------------------------------------------------------------
-
-    fn enum_with_pragma_header() -> String {
-        "
-        {attribute 'qualified_only'}
-        {attribute 'strict'}
-        TYPE E_Color :
-            (Red, Green, Blue);
-        END_TYPE"
-            .to_owned()
-    }
-
-    #[test]
-    fn parse_program_when_pragma_header_and_codesys_dialect_then_ok() {
-        let source = enum_with_pragma_header();
-        let options = CompilerOptions::from_dialect(Dialect::Codesys);
-
-        let result = parse_program(&source, &FileId::default(), &options);
-
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_program_when_pragma_header_and_default_dialect_then_err() {
-        let source = enum_with_pragma_header();
-
-        let result = parse_program(&source, &FileId::default(), &CompilerOptions::default());
-
-        assert!(
-            result.is_err(),
-            "pragmas should still be unrecognized syntax without allow_pragmas"
-        );
-    }
-
-    #[test]
-    fn parse_program_when_pragma_between_declarations_then_ok() {
-        let source = "
-        TYPE E_Color :
-            (Red, Green, Blue);
-        END_TYPE
-        {attribute 'qualified_only'}
-        FUNCTION_BLOCK FB_Example
-        VAR
-            x : INT;
-        END_VAR
-        END_FUNCTION_BLOCK";
-        let options = CompilerOptions::from_dialect(Dialect::Codesys);
-
-        let result = parse_program(source, &FileId::default(), &options);
-
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_program_when_unclosed_pragma_and_codesys_dialect_then_err() {
-        let source = "
-        {attribute 'qualified_only'
-        TYPE E_Color :
-            (Red, Green, Blue);
-        END_TYPE";
-        let options = CompilerOptions::from_dialect(Dialect::Codesys);
-
-        let result = parse_program(source, &FileId::default(), &options);
-
-        assert!(result.is_err());
-    }
-
-    // -----------------------------------------------------------------
-    // CASE branch with no statements.
-    // See specs/plans/2026-07-20-twincat-empty-case-branch.md.
-    // -----------------------------------------------------------------
-
-    fn extract_case(library: &Library) -> Case {
-        let element = library
-            .elements
-            .iter()
-            .find(|e| matches!(e, LibraryElementKind::FunctionBlockDeclaration(_)))
-            .expect("expected a FunctionBlockDeclaration");
-        let fb = cast!(element, LibraryElementKind::FunctionBlockDeclaration);
-        let stmts = cast!(&fb.body, FunctionBlockBodyKind::Statements);
-        cast!(&stmts.body[0], StmtKind::Case).clone()
-    }
-
-    #[test]
-    fn parse_when_case_branch_empty_and_followed_by_another_label_then_ok() {
-        // An empty CASE branch isn't strict IEC 61131-3 (the standard only
-        // allows an explicit empty statement, `5: ;`) -- this is the
-        // `--allow-missing-semicolon` vendor extension filling in the
-        // dropped `;`, so it must be gated behind that flag.
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    x : INT;
-    y : INT;
-END_VAR
-CASE x OF
-    1: y := 1;
-    5: (* no statement here, falls through to nothing *)
-    10: y := 3;
-END_CASE;
-END_FUNCTION_BLOCK";
-        let library = parse_program(source, &FileId::default(), &with_missing_semicolon_flag())
-            .expect("empty CASE branch followed by another label must parse");
-        let case = extract_case(&library);
-        assert_eq!(case.statement_groups.len(), 3);
-        assert!(case.statement_groups[1].statements.is_empty());
-    }
-
-    #[test]
-    fn parse_when_case_branch_empty_and_last_before_end_case_then_ok() {
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    x : INT;
-    y : INT;
-END_VAR
-CASE x OF
-    1: y := 1;
-    5: (* no statement here *)
-END_CASE;
-END_FUNCTION_BLOCK";
-        let library = parse_program(source, &FileId::default(), &with_missing_semicolon_flag())
-            .expect("empty CASE branch as the last one must parse");
-        let case = extract_case(&library);
-        assert_eq!(case.statement_groups.len(), 2);
-        assert!(case.statement_groups[1].statements.is_empty());
-    }
-
-    #[test]
-    fn parse_when_case_branch_empty_and_flag_not_set_then_err() {
-        // Strict IEC 61131-3 (the default dialect) must still reject this --
-        // only `--allow-missing-semicolon` fills in the dropped `;`.
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    x : INT;
-    y : INT;
-END_VAR
-CASE x OF
-    1: y := 1;
-    5: (* no statement here *)
-END_CASE;
-END_FUNCTION_BLOCK";
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_when_case_branch_has_statement_then_regression_ok() {
-        // Regression: an ordinary populated CASE branch must be unaffected.
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    x : INT;
-    y : INT;
-END_VAR
-CASE x OF
-    1: y := 1;
-    5: y := 2;
-END_CASE;
-END_FUNCTION_BLOCK";
-        let library = parse_program(source, &FileId::default(), &CompilerOptions::default())
-            .expect("populated CASE branch must still parse");
-        let case = extract_case(&library);
-        assert_eq!(case.statement_groups.len(), 2);
-        assert_eq!(case.statement_groups[1].statements.len(), 1);
-    }
-
-    // -----------------------------------------------------------------
-    // AND_THEN short-circuit boolean operator.
-    // See specs/plans/2026-07-20-twincat-and-then-operator.md.
-    // -----------------------------------------------------------------
-
-    fn opts_with_short_circuit_operators() -> CompilerOptions {
-        CompilerOptions {
-            allow_short_circuit_operators: true,
-            ..CompilerOptions::default()
-        }
-    }
-
-    fn extract_assignment_value(library: &Library) -> Expr {
-        let element = library
-            .elements
-            .iter()
-            .find(|e| matches!(e, LibraryElementKind::FunctionBlockDeclaration(_)))
-            .expect("expected a FunctionBlockDeclaration");
-        let fb = cast!(element, LibraryElementKind::FunctionBlockDeclaration);
-        let stmts = cast!(&fb.body, FunctionBlockBodyKind::Statements);
-        let assignment = cast!(&stmts.body[0], StmtKind::Assignment);
-        assignment.value.clone()
-    }
-
-    #[test]
-    fn parse_when_and_then_then_ok_and_compare_op_and_then() {
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    a : BOOL;
-    b : BOOL;
-    result : BOOL;
-END_VAR
-result := a AND_THEN b;
-END_FUNCTION_BLOCK";
-        let library = parse_program(
-            source,
-            &FileId::default(),
-            &opts_with_short_circuit_operators(),
-        )
-        .unwrap();
-        let value = extract_assignment_value(&library);
-        let compare = cast!(&value.kind, ExprKind::Compare);
-        assert_eq!(compare.op, CompareOp::AndThen);
-    }
-
-    #[test]
-    fn parse_when_and_then_real_world_shape_then_ok() {
-        // The real motivating shape: guarding a dereference behind a
-        // null-pointer check.
-        let source = "
-FUNCTION_BLOCK FB_Example
-VAR
-    ptr : REF_TO INT;
-    result : BOOL;
-END_VAR
-result := ptr <> 0 AND_THEN ptr^ = 99;
-END_FUNCTION_BLOCK";
-        let options = CompilerOptions {
-            allow_short_circuit_operators: true,
-            allow_ref_to: true,
-            ..CompilerOptions::default()
-        };
-        let result = parse_program(source, &FileId::default(), &options);
-        assert!(result.is_ok(), "parse failed: {:?}", result.err());
-    }
-
-    #[test]
-    fn parse_when_and_then_and_disabled_then_parses_as_identifiers() {
-        // AND_THEN demotes to an ordinary identifier when the flag is
-        // off, matching the pattern used for every other vendor-extension
-        // keyword.
-        let source = "
-FUNCTION_BLOCK FB_ALL_AND_THEN_AS_VAR
-VAR
-    AND_THEN : INT;
-END_VAR
-AND_THEN := 1;
-END_FUNCTION_BLOCK";
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(
-            result.is_ok(),
-            "AND_THEN must remain a valid identifier in standard mode: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn parse_when_enum_all_members_explicit_value_then_parses() {
-        // Matches real usage found in a private test corpus:
-        // TYPE E_ModeLanguage :(Deutsch:=1,English:=2);
-        let source = "
-TYPE E_ModeLanguage : (Deutsch := 1, English := 2);
-END_TYPE";
-        let library = parse_text(source);
-
-        let dt = cast!(
-            &library.elements[0],
-            LibraryElementKind::DataTypeDeclaration
-        );
-        let decl = cast!(dt, DataTypeDeclarationKind::Enumeration);
-        let values = cast!(&decl.spec_init.spec, SpecificationKind::Inline);
-        assert_eq!(values.values.len(), 2);
-        assert_eq!(
-            values.values[0].explicit_value.as_ref().map(|v| v.to_i64()),
-            Some(1)
-        );
-        assert_eq!(
-            values.values[1].explicit_value.as_ref().map(|v| v.to_i64()),
-            Some(2)
-        );
-        assert!(decl.spec_init.underlying_type.is_none());
-    }
-
-    #[test]
-    fn parse_when_enum_first_member_explicit_value_then_parses() {
-        // Matches real cross-repo usage (TcUnit's E_AssertionType, etc.):
-        // only the first member has an explicit value, the rest are
-        // implicit (expected to continue from the previous value + 1).
-        let source = "
-TYPE E_AssertionType :
-(
-    Type_UNDEFINED := 0,
-    Type_ANY,
-    Type_BOOL
-);
-END_TYPE";
-        let library = parse_text(source);
-
-        let dt = cast!(
-            &library.elements[0],
-            LibraryElementKind::DataTypeDeclaration
-        );
-        let decl = cast!(dt, DataTypeDeclarationKind::Enumeration);
-        let values = cast!(&decl.spec_init.spec, SpecificationKind::Inline);
-        assert_eq!(values.values.len(), 3);
-        assert_eq!(
-            values.values[0].explicit_value.as_ref().map(|v| v.to_i64()),
-            Some(0)
-        );
-        assert!(values.values[1].explicit_value.is_none());
-        assert!(values.values[2].explicit_value.is_none());
-    }
-
-    #[test]
-    fn parse_when_enum_base_type_suffix_then_parses() {
-        // Matches real usage (TcUnit's E_AssertionType/E_XmlError): a
-        // CODESYS/TwinCAT base-type suffix after the value list.
-        let source = "
-TYPE E_AssertionType :
-(
-    Type_UNDEFINED := 0,
-    Type_ANY,
-    Type_BOOL
-) BYTE;
-END_TYPE";
-        let library = parse_text(source);
-
-        let dt = cast!(
-            &library.elements[0],
-            LibraryElementKind::DataTypeDeclaration
-        );
-        let decl = cast!(dt, DataTypeDeclarationKind::Enumeration);
-        assert_eq!(
-            decl.spec_init.underlying_type,
-            Some(dsl::common::ElementaryTypeName::BYTE)
-        );
-    }
-
-    #[test]
-    fn parse_when_enum_no_explicit_values_or_base_type_then_parses_unchanged() {
-        // Regression: an ordinary enum declaration must be unaffected.
-        let source = "
-TYPE COLOR : (RED, GREEN, BLUE) := RED;
-END_TYPE";
-        let library = parse_text(source);
-
-        let dt = cast!(
-            &library.elements[0],
-            LibraryElementKind::DataTypeDeclaration
-        );
-        let decl = cast!(dt, DataTypeDeclarationKind::Enumeration);
-        let values = cast!(&decl.spec_init.spec, SpecificationKind::Inline);
-        assert_eq!(values.values.len(), 3);
-        assert!(values.values.iter().all(|v| v.explicit_value.is_none()));
-        assert!(decl.spec_init.underlying_type.is_none());
-        assert_eq!(
-            decl.spec_init.default.as_ref().map(|d| d.value.to_string()),
-            Some("RED".to_string())
-        );
-    }
-
-    // ---------------------------------------------------------------------
-    // Constant-expression VAR initializers.
-    // See specs/plans/2026-07-19-twincat-var-initializer-expressions.md.
-    // ---------------------------------------------------------------------
-
-    fn opts_with_constant_initializer_expressions() -> CompilerOptions {
-        CompilerOptions {
-            allow_constant_initializer_expressions: true,
-            ..CompilerOptions::default()
-        }
-    }
-
-    #[test]
-    fn parse_when_negative_integer_initializer_then_parses_as_simple_not_simple_expr() {
-        // Regression test: switching the initializer grammar from
-        // constant() to expression() must not turn ordinary negative
-        // literals into the vendor-extension SimpleExpr shape, since
-        // expression() routes a leading '-' through its own unary-operator
-        // handling rather than constant()'s built-in signed-literal
-        // parsing. This must parse identically with or without the flag.
-        let source = "PROGRAM main VAR x : INT := -123; END_VAR END_PROGRAM";
-        let library = parse_program(source, &FileId::default(), &CompilerOptions::default())
-            .expect(
-            "negative literal initializers must not require allow_constant_initializer_expressions",
-        );
-
-        let program = cast!(&library.elements[0], LibraryElementKind::ProgramDeclaration);
-        let simple = cast!(
-            &program.variables[0].initializer,
-            InitialValueAssignmentKind::Simple
-        );
-        let lit = cast!(
-            simple.initial_value.as_ref().unwrap(),
-            ConstantKind::IntegerLiteral
-        );
-        assert!(lit.value.is_neg);
-        assert_eq!(lit.value.value.value, 123);
-    }
-
-    #[test]
-    fn parse_when_negative_real_initializer_then_parses_as_simple() {
-        let source = "PROGRAM main VAR x : LREAL := -3.5; END_VAR END_PROGRAM";
-        let library = parse_program(source, &FileId::default(), &CompilerOptions::default())
-            .expect(
-            "negative real initializers must not require allow_constant_initializer_expressions",
-        );
-
-        let program = cast!(&library.elements[0], LibraryElementKind::ProgramDeclaration);
-        let simple = cast!(
-            &program.variables[0].initializer,
-            InitialValueAssignmentKind::Simple
-        );
-        let lit = cast!(
-            simple.initial_value.as_ref().unwrap(),
-            ConstantKind::RealLiteral
-        );
-        assert_eq!(lit.value, -3.5);
-    }
-
-    #[test]
-    fn parse_when_var_enum_default_then_still_resolves_as_enumerated_not_simple_expr() {
-        // Regression test: identifier := identifier is ambiguous between a
-        // simple-typed constant-expression initializer (referencing a
-        // variable) and an enum-typed variable with an enum-value default.
-        // The enum interpretation must still win, matching pre-existing
-        // disambiguation behavior.
-        let source = "
-TYPE
-    MotorState : (STOPPED, RUNNING);
-END_TYPE
-FUNCTION_BLOCK fb1
-VAR
-    State : MotorState := STOPPED;
-END_VAR
-END_FUNCTION_BLOCK";
-        let library = parse_program(
-            source,
-            &FileId::default(),
-            &opts_with_constant_initializer_expressions(),
-        )
-        .unwrap();
-
-        let fb = cast!(
-            &library.elements[1],
-            LibraryElementKind::FunctionBlockDeclaration
-        );
-        assert!(matches!(
-            &fb.variables[0].initializer,
-            InitialValueAssignmentKind::EnumeratedType(_)
-        ));
-    }
-
-    #[test]
-    fn parse_when_arithmetic_initializer_and_flag_disabled_then_still_parses() {
-        // Grammar accepts the broader form unconditionally (parse
-        // permissively, validate later) — only the semantic fold pass
-        // diagnoses a disabled flag, not the parser.
-        let source = "PROGRAM main VAR d2r : LREAL := 4.25/180.0; END_VAR END_PROGRAM";
-        let result = parse_program(source, &FileId::default(), &CompilerOptions::default());
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn parse_when_arithmetic_initializer_then_parses_as_simple_expr() {
-        let source = "PROGRAM main VAR d2r : LREAL := 4.25/180.0; END_VAR END_PROGRAM";
-        let library = parse_program(
-            source,
-            &FileId::default(),
-            &opts_with_constant_initializer_expressions(),
-        )
-        .unwrap();
-
-        let program = cast!(&library.elements[0], LibraryElementKind::ProgramDeclaration);
-        assert!(matches!(
-            &program.variables[0].initializer,
-            InitialValueAssignmentKind::SimpleExpr(_)
-        ));
-    }
-}
